Vegas Baby
- Decreased required database connection (some values now are cached).
- Added the possibility to reactivate an already archived car.
- Chart and map resources only will be loaded when neccessary.
