+++ MileIT/WebContent/WEB-INF/pages/includes/activateModal.jsp
@@ -1,0 +1,27 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
+<%@ taglib prefix="fmt" uri="http://java.sun.com/jsp/jstl/fmt"%>
+
+<c:set var="lang" value="${not empty param.lang ? param.lang : not empty lang ? lang : pageContext.request.locale}" />
+
+<fmt:setLocale value="${lang}" />
+<fmt:setBundle basename="hu.thom.mileit.i18n.mileit" />
+
+<div class="modal fade" id="confirm-activate" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
+	<div class="modal-dialog">
+		<div class="modal-content">
+			<div class="modal-header">
+				<fmt:message key="activate.modal.title" />
+			</div>
+			<div class="modal-body">
+				<fmt:message key="activate.modal.body" />
+			</div>
+			<div class="modal-footer">
+				<button type="button" class="btn btn-default" data-dismiss="modal">
+					<fmt:message key="button.cancel" />
+				</button>
+				<a class="btn btn-success btn-ok"><fmt:message key="button.activate" /></a>
+			</div>
+		</div>
+	</div>
+</div>
