+++ MileIT/src/hu/thom/mileit/i18n/mileit.properties
@@ -15,7 +15,7 @@ nodata=There is no data found.
 # Status messages
 status.new=You have successfully created a new item!
 status.edit=You have successfully modified the item!
-status.archive=You have successfully archived the selected item!
+status.archive=You have successfully archived or reactivated the selected item!
 status.delete=You have successfully deleted the selected item!
 status.missing=The highlighted fields are mandatory!
 status.error=An application error occured. Please try again.
@@ -55,6 +55,7 @@ table.actions=Actions
 button.new=New
 button.save=Save
 button.archive=Archive
+button.activate=Reactivate
 button.cancel=Cancel
 button.map=Add to car
 button.delete=Delete
@@ -62,14 +63,20 @@ button.edit=Edit
 
 # Archive modal
 archive.modal.title=Confirm archive
-archive.modal.body=Are you sure to archive the selected item? This action cannot be undone!
+archive.modal.body=Are you sure to archive the selected item?
 
 # Delete modal
 delete.modal.title=Confirm delete
 delete.modal.body=Are you sure to delete the selected item? This action cannot be undone!
 
+# Activate modal
+activate.modal.title=Confirm reactivation
+activate.modal.body=Are you sure to reactivate the selected item?
+
 # Manage cars
 cars.cars.title=Manage cars
+cars.cars.title.active=Active cars
+cars.cars.title.archived=Archived cars
 cars.cars.title.new=Add new car
 cars.cars.title.edit=Edit car
 cars.form.manufacturer=Manufacturer
